Create a more readable help target (#51)
The help target shows sections of the target. Colorization for targets
in the output helps to recognize targets.

Remove the `./` prefix from targets where a path needs to be specified.
This causes some target not properly rendered in the help output.

Remove `vendor` as a dependency and it is already part of the source tree.