[ODC-1240] More Unit and E2E Testing (#55)
Adding more unit and E2E tests coverage to the project, and going back on previous PRs to make sure the scheme is adopted in the project. Previously, the operator has been making use of a special label and on the later scheme is replaced. Other changes are also part of this PR and described below.