Make CatchAllReconciler implement Reconcile again
As a quick hack, the service binding request selector value is mocked
in the object's name since there are some limitations when using the
Reconcile interface. Basically, since it offers a concrete
reconcile.Request type, there's no way to augment its surface to add
extra data to be carried to workers.