Add note about test
As the comment says, I wasn't very sure what the tested wanted to
express so I just inferred an error shouldn't be returned in case
gibberish is fed (this behavior was implemented in another commit),
and changed the test to assert error shouldn't be returned.

Since the test seems not to be clear, I just made it pass for now.