Add ServiceBinder
ServiceBinder is a component to orchestrate the lower level
components when binding or unbinding an application from its
dependencies.

This change doesn't yet add it to the reconciler, just bring
the changes in such a way the project compiles and the existing
tests still pass.

Tests will appear in a different commit.