pre-req refactor to support multiple backing services (#293)
* Fixes go mod invalid pseudo-version error

With the current version specified in go.mod, the following
error happens:

$ go mod vendor
go: github.com/openshift/api@v3.9.1-0.20190424152011-77b8897ec79a+incompatible: invalid pseudo-version: preceding tag (v3.9.0) not found

This patch replaces it with a valid pseudo-version encoding
the same commit specified in the original.

* Once more fix operator-sdk identifier

* Add RelatedResources type

A RelatedResource is currently a CR and CRDDescription pair of a resource
related to a Service Binding Request.

* Add methods and stubs that will be used by the ServiceBinder

* Add Plan methods that will be used by ServiceBinder

* Add BackingServiceSelectors field in ServiceBindingRequest

* Flattens GVK information accordingly

* fixup! Add methods and stubs that will be used by the ServiceBinder

* Changes Secret.Get() signature for ServiceBinder

* Add ServiceBinder

ServiceBinder is a component to orchestrate the lower level
components when binding or unbinding an application from its
dependencies.

This change doesn't yet add it to the reconciler, just bring
the changes in such a way the project compiles and the existing
tests still pass.

Tests will appear in a different commit.

* Add ServiceBinder tests

Tests are still being skipped since it is incomplete.

* Revert "Once more fix operator-sdk identifier"

This reverts commit 0d77841c

* Adjust left overs from merge

* Fix lint error (gosimple)

* Fix lint error (ineffassign)