Adjust symbols visibility throughout the system (#465)
This change adjusts the visibility of symbols in the entire project; the
visibility has been raised several moments during review of #407 and
tracked in #462 after #407 has been merged.

Because of the nature of the change, some variables have been renamed to
avoid value shadowing when the variable name is identical to the type
name (e.g. 'binder := &binder{}').

Closes #462