Fix bug where binding data is expected to overwrite the related resource name (#515)
* refactor(annotations): :recycle: add RawData in result

* refactor(mocks): :white_check_mark: adjust test fixtures

* fix(servicectx): :bug: ignore errors other than invalid annotation prefix and handler not found

This change modifies buildServiceContexts() to ignore certain errors
likely to happen, such as an invalid annotation format has been fed or
in the case the format is proper, whether there's a handler for the
given annotation format.

* refactor(servicectx): extract runHandler() from buildServiceContext()

* refactor(mocks): add data argument to SecretMock

* refactor(reconciler): sort selectors by name and namespace for predictability

* refactor(servicectx): make tests more explicit

* docs(annotations): add docs for RawData

* refactor(tests): fix environment name name and annotation

* refactor(tests): resources related to a services must exist before the service

* refactor: add SKIP_CLEANUP_ERROR env variable to be overriden in local testing situations

* refactor(tests): disable unsupported scenarios

Scenarios where the SBR is created prior the application and/or service
are not well supported yet. Currently annotations added to either
service and application are used to correlate those resources, but those
are added by the SBO when the SBR is reconciled; without this
information currently it is not possible to correlate the resource to a
SBR.