Fix: Remove volume and volumemount from app resource when unbinding (#950)
When projecting bindings as files, we inject volume and volume mount carrying the same name
as binding resource, but when binding is removed, we wrongly look to remove volume and volume mount
having the name equal to the binding secret.

This change fixes it and introduces the missing acceptance tests.

Signed-off-by: Predrag Knezevic <pknezevi@redhat.com>