Regression fix: refer multiple services in custom mapping templates
service structs are initialized inside a `for range` loop, and the loop variable is passed as a reference to each of service structs.
However, with each new iteration its value changes, but the reference remains the same, leading that the value of the last iteration
is passed to all service structs eventually.

* do not pass loop variable to service structs
* unit tests were extended to detect regressions in the future
* `Bind two backend services by creating 1 SBR to a single application` acceptance test was modified to detect regressions in the future

Signed-off-by: Predrag Knezevic <pknezevi@redhat.com>