fix compile error when using SBO as library with newer controller-runtime (#1039)
SBO uses currently controller-runtime 0.6.4 and there the admission webhook `Request` struct
uses v1beta1 API. Later controller-runtime versions use `v1` version, and this
turned to generate compile error when SBO is included as dependency in a project
using latest controller-runtime as well.

The root cause for the issue is that the admission webhook is part of `api` package,
and therefore needs to be complied even if not used in the project.

There is no reasons why the webhook code need to be a part of `api` package, and
the issue is fixed by moving it in separate `apis/webhooks` package.

Two new test projects are added so that we can catch regressions.

Signed-off-by: Predrag Knezevic <pknezevi@redhat.com>