Add support for CrunchyData Postgres operator
In order to read binding data from the right secret, the value of `path` key of a service binding annotation
is treated as a template:
* multiple json paths can specified and the results of their evaluation are rendered in the final binding value
* the template can contain arbitrary additional strings that are rendered 'as-is' in the final binding value

Hence, it is possible to generate proper secret reference and read binding data from it. In the case of
CrunchyData Postgres Operator, this is the right annotations:

```
"service.binding": "path={.metadata.name}-pguser-{.metadata.name},objectType=Secret",
```

On top of that, `user` is exposed as `username` as well, and `dbname` and `database`, so that
more applications can find all the binding data they need.

Signed-off-by: Predrag Knezevic <pknezevi@redhat.com>