Add keycloak sample for spring-petclinic app
This sample provides a binding of the spring-petclinic application to a backing Keycloak as an Identity Provider. After the successful projection of the binding values the application will be auto-configured for user authentication via the OpenId Connect authorization code flow. The sample includes a deployment for the Keycloak server that creates a realm 'Test' on startup. Users need to be created manually. The sample also includes an ingess definition for the Keycloak server. The ingress MUST be DNS resolvable with 'keycloak.local' for the sample to work.

Since the petclinic application has been extended to utilize the spring-boot-starter-oauth2-client that in turn includes Spring Security the auto-configuration in absence of the binding values i.e. before the Keycloak or the Service Binding is deployed, will result in a random password being generated on startup. This password will be written to the log and provides an alternative means of authentication for the also auto-generated user 'user'. This affects all other sample that use the extended petclinic application.

Related issue: redhat-developer#1083

Signed-off-by: Björn Großewinkelmann <github@ankertaugasse.de>