This is our PostgreSQL Service Endpoint Definition (SED)
This SED can be used in cases where the service was provisioned
either manually or via a Helm Chart. It will allow developer to
test their binding data in a standard maner regradless of how the
PostgreSQL database was providioned or the backend cloud. The
assumption here is that the is no CR representing the service in
kubernetes.