ci: get CI working again (#1127)
In commit ea79796cb ("update to operator-sdk 1.3.0 (#848)"), we upgraded
controller-gen to 0.8.0.  However, v0.8.0 requires go 1.17 or greater,
which is currently unavailable in Red Hat's container registries (i.e.
there's no container on catalog.redhat.com); consequently, SBO is built
with go 1.16.  Until that is rectified, downgrade controller-gen to
0.7.0, which builds with go 1.16.

At the same time, use `go install` instead of the deprecated `go get`
when we install local build tools, such as controller-gen, yq, and
kustomize.  This requires a version bump of kustomize to work correctly
as well as adding an extra step to install a recent version of go in the
non-OLM github actions acceptance test runner.

Finally, with the release of go 1.18, our go linting needs to become
aware of this new release, since our linting runs with the latest
release of go.  This can be fixed by updating the golangci/golangci-lint
runner to v1.45.2; since v1.45.1, it will autodetect the language
version of a project.

Signed-off-by: Andy Sadler <ansadler@redhat.com>